+++ src/noahtheduke/splint/parser/defn.clj
@@ -2,7 +2,7 @@
 ; Clojure: Copyright (c) Rich Hickey. All rights reserved., ELP 1.0
 ; Modifications licensed under ELP 1.0
 
-(ns noahtheduke.spat.parser.defn
+(ns noahtheduke.splint.parser.defn
   (:require
     [noahtheduke.spat.pattern :refer [drop-quote]]))